Split lex into the plain logos part, and a token wrapper things.
