+++ src/test_util.rs
@@ -1,6 +1,6 @@
 use crate::codespan;
 use crate::error::*;
-use crate::lex;
+use crate::token_wrap::*;
 use crate::parser;
 use codespan_reporting::term;
 use codespan_reporting::term::termcolor::{ColorChoice, StandardStream};
@@ -12,7 +12,7 @@ pub fn do_test<'a>(sources: &[&'a str]) -> Result<(), Vec<(&'a str, Error<'a>)>>
         .map(|(index, s)| {
             (
                 index,
-                parser::propParser::new().parse(lex::Tokens::from_string(s)),
+                parser::propParser::new().parse(Tokens::from_string(s)),
             )
         })
         .partition(|(_, r)| r.is_ok());
